pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG Chatbot using Gemini API + FAISS + LangChain"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.2.0",
    "faiss-cpu>=1.8.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "streamlit>=1.35.0",
    "tiktoken>=0.7.0",
    "langchain-text-splitters>=1.1.1",
    "langchain-classic>=1.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
